backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6

sqlalchemy==2.0.36
pydantic==2.9.2
pydantic-settings==2.5.2

python-dotenv==1.0.1
python-multipart==0.0.9

httpx==0.27.2

python-jose==3.3.0
passlib[bcrypt]==1.7.4

reportlab==4.2.2
jinja2==3.1.4

orjson==3.10.7
